Replace JSON and XML strategy with SvcLoaderStrategy
This new strategy is lazy loaded, which means the choose method is not
called during app bootstrap, but only the first time a MediaType is
invoked.
Also, the CACHE key is a instance of ProviderStrategy and not the
class name anymore.
